pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livekit-ai-switchboard"
version = "0.0.1"
description = "Intelligent LLM routing for LiveKit voice agents"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{name = "vantxlab"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["livekit", "llm", "routing", "voice", "agents"]
dependencies = [
    "livekit-agents>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "python-dotenv>=1.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/vantxlab/livekit-ai-switchboard"
Repository = "https://github.com/vantxlab/livekit-ai-switchboard"
Issues = "https://github.com/vantxlab/livekit-ai-switchboard/issues"

[tool.hatch.build.targets.wheel]
packages = ["ai_switchboard"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires API keys and makes real LLM calls",
]
